Refactor project loading logic to projectService and add SYSTEM_MAP.md
- Created `docs/SYSTEM_MAP.md` to document system architecture, routes, components, and data flow.
- Created `frontend/src/domain/projectService.js` to handle project data loading, normalization, and default value merging.
- Refactored `frontend/src/store/projectSlice.js` to use `projectService.js` instead of embedding business logic directly.
- Updated `docs/CONTRIBUTING.md` to reference the new system map.
- Updated internal imports in `assetService.js` and `projectService.js` to use explicit `.js` extensions for ESM compatibility.
