+++ frontend/src/store/projectSlice.js
@@ -1,8 +1,6 @@
 import { API } from '../lib/api';
-import { syncAssetColors, forkAsset } from '../domain/assetService';
-import { normalizeAsset } from '../lib/utils'; // Import normalize helper
-
-const DEFAULT_COLORS = { room: '#fdfcdc', furniture: '#8fbc8f', fixture: '#cccccc' };
+import { syncAssetColors } from '../domain/assetService';
+import { loadProjectData, DEFAULT_COLORS } from '../domain/projectService';
 
 export const createProjectSlice = (set, get) => ({
     projects: [],
@@ -20,56 +18,12 @@ export const createProjectSlice = (set, get) => ({
 
         set({ currentProjectId: projectId });
 
-        const [projectData, globalAssetsData, paletteData] = await Promise.all([
-            API.getProjectData(projectId),
-            API.getAssets(),
-            API.getPalette()
-        ]);
+        const newState = await loadProjectData(projectId, API);
 
-        // Normalize global assets to use entities
-        const globalAssets = (globalAssetsData || []).map(a => {
-            const normalized = normalizeAsset(a); // Helper from utils
-            return { ...normalized, source: 'global' };
-        });
-
-        const colorPalette = paletteData?.colors || [];
-        const globalDefaultColors = paletteData?.defaults || DEFAULT_COLORS;
-        const categoryLabels = paletteData?.labels || {};
-        const projectDefaultColors = projectData?.defaultColors || {};
-
-        // Merge defaults: Project specific overrides global
-        const defaultColors = { ...globalDefaultColors, ...projectDefaultColors };
-
-        // Normalize loaded local assets
-        let loadedAssets = (projectData?.assets || []).map(normalizeAsset);
-        let instances = projectData?.instances || [];
-
-        // Logic: Fork Global Assets if Project Empty
-        if (loadedAssets.length === 0) {
-            loadedAssets = globalAssets.map(ga => forkAsset(ga, defaultColors));
-        } else {
-            // Sync existing local assets
-            // Note: syncAssetColors also handles basic normalization if missed
-            loadedAssets = syncAssetColors(loadedAssets, defaultColors);
+        if (newState) {
+            set(newState);
+            get().temporal?.clear();
         }
-
-        set({
-            globalAssets,
-            colorPalette,
-            globalDefaultColors,   // Store base global defaults
-            projectDefaultColors,  // Store project-specific overrides
-            defaultColors,         // Effective defaults
-            categoryLabels,
-            localAssets: loadedAssets,
-            instances,
-            selectedIds: [],
-            designTargetId: null,
-            selectedShapeIndices: [],
-            selectedPointIndex: null,
-            viewState: { x: 50, y: 600, scale: 1 }
-        });
-
-        get().temporal?.clear();
     },
 
     saveProjectData: async () => {
